Feito a translação, a escala (sem pegar o centro) e a rotação (sem pegar o centro), a criação de objeto, a mudança de primitiva, a deleção de objeto, mais keymappings, a camera, o pan, e mais algumas coisas

+++ BBox.cs
@@ -2,12 +2,12 @@
 
 namespace TrabalhoCG3 {
     public class BBox {
-        public float MinX { get; set; }
-        public float MinY { get; set; }
-        public float MinZ { get; set; }
-        public float MaxX { get; set; }
-        public float MaxY { get; set; }
-        public float MaxZ { get; set; }
+        public double MinX { get; set; }
+        public double MinY { get; set; }
+        public double MinZ { get; set; }
+        public double MaxX { get; set; }
+        public double MaxY { get; set; }
+        public double MaxZ { get; set; }
 
         public Point4D Center { 
             get {
